+++ ui/components/RequestsLevel.vue
@@ -3,7 +3,7 @@
         <div class="level-item has-text-centered">
             <div>
                 <p class="has-tooltip-arrow has-tooltip-top has-tooltip-multiline" data-tooltip="Requests that returned a 100, 200, or 300 level response code.">
-                    <span class="heading">Successful</span>
+                    <span class="heading"><span class="icon"><i class="fas fa-check mr-3"></i></span>Successful</span>
                 </p>
                 <p class="title">{{ requests.successful.toLocaleString() }}</p>
             </div>
@@ -11,7 +11,7 @@
         <div class="level-item has-text-centered">
             <div>
                 <p class="has-tooltip-arrow has-tooltip-top has-tooltip-multiline" data-tooltip="Requests that were rejected due to a client error (4xx).">
-                    <span class="heading">Rejected</span>
+                    <span class="heading"><span class="icon"><i class="fas fa-exclamation mr-3"></i></span>Rejected</span>
                 </p>
                 <p class="title">{{ requests.rejected.toLocaleString() }}</p>
             </div>
@@ -19,7 +19,7 @@
         <div class="level-item has-text-centered">
             <div>
                 <p class="has-tooltip-arrow has-tooltip-top has-tooltip-multiline" data-tooltip="Requests that failed due to a server error (5xx).">
-                    <span class="heading">Failed</span>
+                    <span class="heading"><span class="icon"><i class="fas fa-times mr-3"></i></span>Failed</span>
                 </p>
                 <p class="title">{{ requests.failed.toLocaleString() }}</p>
             </div>